refactor: adjust config for alerts
