+++ src/config/demoGuideConfig.js
@@ -264,3 +264,15 @@ export const styles = {
     color: 'black',
   }),
 };
+
+// Alerts
+export const isAlertOpen = atom({
+  key: 'isAlertOpen', // unique ID (with respect to other atoms/selectors)
+  default: false, // default value (aka initial value)
+});
+
+// Alert content to display
+export const alertContent = atom({
+  key: 'alertContent', // unique ID (with respect to other atoms/selectors)
+  default: '', // default value (aka initial value)
+});
